Issue 138:	Calendar export does not include Isha Issue 139:	Calendar exporting may not work properly

+++ SalatTen/src/applicationui.hpp
@@ -63,7 +63,7 @@ public:
     virtual ~ApplicationUI();
 
     Q_INVOKABLE void loadAccounts();
-    Q_INVOKABLE void exportToCalendar(int numDays, QVariantList const& events, int accountId);
+    Q_INVOKABLE void exportToCalendar(int numDays, QVariantList const& events, qint64 accountId);
     Q_INVOKABLE void renderMap(QObject* mapView, qreal latitude, qreal longitude, QString const& name, QString const& event, bool focus=false);
     Q_INVOKABLE QObject* refreshLocation();
     Q_INVOKABLE void cleanupCalendarEvents();
